Bug 1767590 - Initialize pointers passed to NSS_CMSDigestContext_FinishMultiple r=nss-reviewers,jschanck
As NSS_CMSDigestContext_FinishMultiple may leave its outparam unchanged when it returns SECSuccess, ensure
that we set the value to NULL prior to invoking it. If this has happened because data was missing and hence the
digest was never updated, the secasn1d parser will notice the missing child and raise a decodeError.

Differential Revision: https://phabricator.services.mozilla.com/D145425