Bug 1784191 - Cast the result of GetProcAddress. r=bbeurdouche,nss-reviewers
clang trunk recently added a check for incompatible function pointers,
and now is not happy with a FARPROC being given as a
PRThreadPrivateDTOR, because that's not the same signature. But FARPROC
is the "normal" return type from GetProcAddress, but OTOH, it's not the
actual type signature for PR_Free.

Differential Revision: https://phabricator.services.mozilla.com/D154337

--HG--
extra : moz-landing-system : lando