Bug 1802331 - compress docker image artifact with zstd. r=nss-reviewers,releng-reviewers,jlorenzo,bbeurdouche
When we moved tasks to run on GCP from AWS in bug 1799315, we started
using a newer version of docker-worker including the changes from bug
1637302; as a result, artifacts are compressed with gzip before upload
to s3, and downloads now come with a "content-encoding: gzip" header and
compressed content, regardless of the client's "accept-encoding".
Unfortunately docker-worker doesn't handle that encoding and expects an
artifact called image.tar to be uncompressed.  To work around that
issue, we now compress docker images in image_builder with zstd before
upload.

[Ideally we'd install the zstd package in the nssdev/image_builder
docker image itself instead of doing it in every task, however I'm not
sure who owns that or how it's built so this might be good enough for
right now.]

Differential Revision: https://phabricator.services.mozilla.com/D163306