Bug 1803226 - NULL password encoding incorrect. r=nss-reviewers,jschanck
The test for adding the unicode null in the null password case was incorrect from Bug 1757075 (https://bugzilla.mozilla.org/show_bug.cgi?id=1757075). The sense of the test was backwards meaning that no null was added.

We didn't notice because NSS and openssl tolerate incorrect null password encoding. It was picked up in gnutls interop testing.

Differential Revision: https://phabricator.services.mozilla.com/D163498