Bug 1804688 - Correct addition of GREASE value to ALPN xtn. r=mt
A long-standing comment suggested the length field should be 1 byte, but the
code was adding a two-byte length. Inspection showed that the length field
should indeed be two-bytes. This patch corrects the comment and adjusts
the length calculation for the ALPN GREASE value. Tests are included in the
child patch of this revision.

Differential Revision: https://phabricator.services.mozilla.com/D169620