Bug 1804660 - Make high tag number assertion failure an error. r=nss-reviewers,djackson
If a template has an OPTIONAL field, and we find that the input does not match
that field's tag number, we mark the field as missing. If the next field is an
ASN.1 ANY, we need to write the previously-parsed tag number out. Since high
tag number forms are rare, we never implemented the necessary re-encoding of
multi-byte tags, and we noted this with an assertion. That assertion is
remotely triggerable in debug builds. This patch removes the assertion and
returns a SEC_ERROR_LIBRARY_FAILURE instead.

Differential Revision: https://phabricator.services.mozilla.com/D170678

--HG--
extra : moz-landing-system : lando