Bug 1774659 NSS needs a database tool that can dump the low level representation of the database. r=jschanck
When debugging the database, it would be helpful to know what is in the database is a nicely formated way. certutil dumps a high level view of the certs and keys, sqlite3 can dump the low level tables and raw entries. It would be useful to dump the database as softoken sees the database.

This code grabs a copy of the latest sdb.c from softoken and uses it to fetch the database entries, then parses them as necessary. It uses the pkcs11 table in libsec to format the result data into human readable strings.

Differential Revision: https://phabricator.services.mozilla.com/D165103