Bug 1896353 Use lib::pkix as default validator instead of the old-one r=rrelyea
It's about changing what cert_VerifyCertChain() in lib/certdb/certvfy.c uses. Currently it defaults to cert_VerifyCertChainOld() instead of cert_VerifyCertChainPkix(). There's a function it calls to determine which is uses called CERT_GetUsePKIXForValidation(). You need to change that function from a default of FALSE (overridden by an environment variable) to a default of TRUE (with a different value for the environment variable setting it back)

Differential Revision: https://phabricator.services.mozilla.com/D210287