Bug 1184059 - mozilla::pkix: allow reference ID labels to begin and/or end with hyphens r=jschanck
Hyphens shouldn't appear at the beginning or end of labels. However, in
practice, sometimes a website will use a wildcard certificate and attempt to
match it with a reference ID with such a label. For compatibility, this change
allows reference ID labels to begin and/or end with hyphens.

Differential Revision: https://phabricator.services.mozilla.com/D230003

--HG--
extra : moz-landing-system : lando