Bug 1933031 Segmentation fault in SECITEM_Hash during pkcs12 processing
The crash now only happens in pk12util if integrity is ignored. The underlying crash, however is the result of errors in the pk12util handling. It is dereferencing pointers that can be null without checking them. This patch fixes the dereference issues as well as making SECOID_GetOID more robust.

There is a question of why errors in the cert bag looks like integrity errors. That may have been an issue with the previous patch for bug 1923767.

Differential Revision: https://phabricator.services.mozilla.com/D231585

--HG--
extra : rebase_source : bdcbfb1172c63cc24332d3720df384c7e8a21b23