Bug 1930806 FIPS changes need to be upstreamed: FIPS 140-3 RNG
For FIPS-140-3 our seed RNG needs to meet certain standards. On Linux this is accomplished by calling getrandom using GRND_RANDOM. This causes us to pull from the /dev/random pool rather than /dev/urandom pool. This drains the entropy in /dev/random counter, so you only want to do this if you are in FIPS mode. We also don't want to fallback onto to the old /dev/urandom devices in this case.

Differential Revision: https://phabricator.services.mozilla.com/D232139