Bug 1939512: Call BL_Init before RNG_RNGInit() so that special SHA instructions can be used if available r=#rrelyea
While looking at something a while back I noticed that some calls to SHA functions at startup weren't using the versions that use the special SHA instructions, even though my machine supported them, but later calls were using these special versions.

A brief investigation showed that this is was because the flag that is checked to determine whether or not to use those instructions is set right after those functions are used. Switching the order of the calls to BL_Init() and RNG_RNGInit() in pkcs11.c fixes this.

Looking around quickly I couldn't see any reason why this needed to be the case, and I looked through the commit history and didn't see any indication that this was done on purpose.

To be clear, this does not make any huge difference, but I do get a consistent and measurable reduction in the time that RNG_RNGInit() takes to run on my machine (0.019ms vs 0.054ms).

Differential Revision: https://phabricator.services.mozilla.com/D232988