Bug 1930806 FIPS changes need to be upstreamed: force ems policy
This patch creates an ems policy which forces the use of ems in ssl for tls 1.0, tls 1.1 and tls 1.2. FIPS requires ems. We enforce FIPS algorithm usage by policy, so we need a policy which allows us to force the use of ems.

The code enforces policy be checking if the policy value is set if we are about to use non ems Master secret generation and fail if the policy is set. The policy is off by default, it can be turned on explicitly. if someone says disable=all in the policy, the policy will still be off.

This check makes sure use of non-ems is prevented in tls (1.0-1.2). It doesn't always lead to pretty error messages. It might be worthwhile adding a check at the end of both the handle client hello and handle server hello  to make sure the ems is negotiated for tls 1.0-1.2 connections if this policy is set. That is not in this patch. Even if we add it, it's good to have the paranoia fallback to make sure we can't accidentally use non-ems master secrets if the policy is turned on.

Differential Revision: https://phabricator.services.mozilla.com/D232245