Bug 1974505 - Key private/public/secret keys by key type in softoken keydb r=nss-reviewers,rrelyea
When attempting to add a new key with the same ID but a different type than an
existing key, handle that as a new key instead of overwriting the existing key.
This is the expected behavior by the changes in Bug 198090.

Additionally, fix a bad strcmp in sdrtest which led to this bug being undetected
in the first place.

Differential Revision: https://phabricator.services.mozilla.com/D255378

--HG--
extra : moz-landing-system : lando