Bug 1983313 make group and scheme support in test tools generic
As part of turning off xyber, and enabling secp256r1mlkem768, secp384r1mlkem1024, and ml-dsa, we would want our test tools to be able to recognize these new groups and schemes on their command line. In fact they will recognize them by just updating secutil.c, but they won't show up in the command line.

This patch turns the secutil.c serial if string compares to a table and a loop with string compares. The table is then also used to convert our grp or scheme number to a string for display, and a function that lets us walk down the list so we can display the list in our help automatically.

This this change, we no longer need to update selfserv, tstclnt, and strclnt when adding (or removing) new groups and signature scheme, we just need to update a table in secutil.c

Differential Revision: https://phabricator.services.mozilla.com/D261766

--HG--
extra : rebase_source : ef2f4caa19c4efd1aea411f32164dc9443ab18c8