Bug 1983320 ml-dsa: Prep Create a FindOidTagByString function
Prep - create a FindOidTagBy description function and replace all the one off in other libraries and tools in NSS.

Most uses are tooking for crypto algorithms, so in that case we'll filter by the existance of a PKCS #11 mechanism. The function can also take an optional len, so we do a strncmp. This allows comparing a substring .

Differential Revision: https://phabricator.services.mozilla.com/D263588