Bug 1983320 ml-dsa: Prep Create a CreateSignatureAlgorithmID function
Create two new public functions:

SEC_CreateSignatureAlgorithmID

Does the full SignatureAlgorithmID creation using the the key and some optional parameters. It replaces code that called SEC_GetSignatureAlgorithmOidTag() and
SEC_SetSignatureAlgorithmID(). The advantage of this function is it can automatically format any parameters needed in the SignatureID. While it's mostly for Signing (and thus using the private key), it can be used with the public key to create a SignAlg for verification.

SEC_GetSignatureAlgorithmOidTagByKey

This replaces SEC_GetSignatureAlgorithmOidTag(). The previous function only took the key type. For ml-dsa and future signing algorithms, we'll need the full key to determine the OidTag.

Also update our tests and internal uses to use these new functions. The existing tests/cert/cert.sh and tests/tools/tools.sh should be sufficient to test these new function.

Differential Revision: https://phabricator.services.mozilla.com/D263657