Bug 1988046 - SEC_ASN1Decode* should ensure it has read as many bytes as each length field indicates r=nss-reviewers,jschanck
More than 25 years ago (i.e., before NSS was open-sourced), an NSS DER encoder
apparently had a bug whereby it would say a value would be longer than it would
turn out to be. The workaround involved making the decoder pretend it had read
the entirety of the value when it had not. In addition to accepting invalid
encodings, this had the side-effect of allowing end-of-contents bytes in
definite length contexts, which is invalid. As a consequence of this, it has
been possible to smuggle data that would never be read or validated. This would
probably only be an issue when verifying a signature that used a weak hash
algorithm.

Differential Revision: https://phabricator.services.mozilla.com/D265022

--HG--
extra : moz-landing-system : lando