Bug 2017997 CKA_PARAM_SET missing from the CK_ULONG list in softoken.
CKA_PARAMETER_SET is a CKA_ULONG, and should be stored in network order in the database, but it's missing from the list. It needs to be added to the list and we need to be able to read incorrectly stored versions (since the bad version has been shipped for a couple of releases now).

Differential Revision: https://phabricator.services.mozilla.com/D284189

--HG--
extra : rebase_source : 6227ac3d544851ed621bb19fb8c92d3bd8913a99